fix: fix check department and routes
